+++ common/src/main/java/com/dudko/blazinghot/data/conditions/DefaultLoadConditions.java
@@ -1,10 +1,12 @@
 package com.dudko.blazinghot.data.conditions;
 
+import java.util.List;
 import java.util.stream.Stream;
 
 import com.dudko.blazinghot.compat.Mods;
 
 import net.minecraft.tags.TagKey;
+import net.minecraft.world.level.ItemLike;
 
 public class DefaultLoadConditions {
 
@@ -16,12 +18,20 @@ public class DefaultLoadConditions {
 		return allModsLoaded(Stream.of(mods).map(m -> m.id).toArray(String[]::new));
 	}
 
+	public static LoadCondition<String> allModsLoaded(List<Mods> mods) {
+		return allModsLoaded(mods.stream().map(m -> m.id).toArray(String[]::new));
+	}
+
 	public static LoadCondition<String> anyModLoaded(String... mods) {
 		return new LoadCondition<>(LoadCondition.Type.ANY_MOD_LOADED, mods);
 	}
 
 	public static LoadCondition<String> anyModLoaded(Mods... mods) {
 		return anyModLoaded(Stream.of(mods).map(m -> m.id).toArray(String[]::new));
+	}
+
+	public static LoadCondition<String> anyModLoaded(List<Mods> mods) {
+		return anyModLoaded(mods.stream().map(m -> m.id).toArray(String[]::new));
 	}
 
 	public static LoadCondition<LoadCondition<?>> or(LoadCondition<?>... conditions) {
@@ -34,6 +44,10 @@ public class DefaultLoadConditions {
 
 	public static LoadCondition<LoadCondition<?>> not(LoadCondition<?> condition) {
 		return new LoadCondition<>(LoadCondition.Type.NOT, condition);
+	}
+
+	public static LoadCondition<ItemLike> itemsRegistered(ItemLike... items) {
+		return new LoadCondition<>(LoadCondition.Type.ITEMS_REGISTERED, items);
 	}
 
 	@SafeVarargs